+++ crates/cubecl-farm/Cargo.toml
@@ -17,7 +17,7 @@ cubecl-runtime = { path = "../cubecl-runtime", version = "0.6.0", default-featur
 ] }
 cubecl-cuda = { path = "../cubecl-cuda", version = "0.6.0" }
 bytemuck = { workspace = true }
-cudarc = { version = "0.16.3", features = [
+cudarc = { version = "0.16.4", features = [
   "driver",
   "nccl",
 ], default-features = false }
