+++ crates/cubecl-farm/src/basic.rs
@@ -141,23 +141,3 @@ pub fn assign_units_proportional(unit_count: usize, proportions: &[f64]) -> Vec<
     }
     assignments
 }
-
-#[derive(Debug)]
-pub enum CudaError {
-    DriverError(cudarc::driver::DriverError),
-    InvalidDevice,
-    InvalidConfiguration,
-    NcclError(cudarc::nccl::result::NcclError),
-}
-
-impl From<cudarc::driver::DriverError> for CudaError {
-    fn from(err: cudarc::driver::DriverError) -> Self {
-        CudaError::DriverError(err)
-    }
-}
-
-impl From<cudarc::nccl::result::NcclError> for CudaError {
-    fn from(err: cudarc::nccl::result::NcclError) -> Self {
-        CudaError::NcclError(err)
-    }
-}
